requirements.txt
numpy==1.24.3
pillow==10.3.0
colorlog==6.8.2
termcolor==2.4.0
tqdm==4.66.2
kinpy==0.2.2
ipython==8.12.3
h5py==3.11.0
open3d==0.16.0
hydra-core==1.3.2
pyserial==3.5
omegaconf==2.3.0
pyrealsense2==2.55.1.6486
opencv-python==4.5.5.64
opencv-contrib-python==4.5.5.64
torch==1.13.0
torchvision==0.14.0
torchaudio==0.13.0
transformers==4.46.3
einops==0.3.0
open_clip_torch==2.0.2
pytorch-lightning==1.5.0
pytorch-kinematics==0.7.5
pynput==1.7.7
six==1.17.0
imageio==2.9.0
scipy==1.10.1
matplotlib==3.7.5
natsort==8.4.0
imageio-ffmpeg==0.4.2
